requirements.txt
# Core dependencies

pandas==2.1.4
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# API Clients
requests==2.31.0
openai==2.15.0
httpx==0.28.1

# Scheduling
APScheduler==3.10.4

# Web Framework (optional)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
